# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sugar Daddy Platform API
  description: |
    Comprehensive API documentation for the Sugar Daddy/Sugar Baby platform.
    This API provides authentication, user management, profile management, 
    matching, messaging, and subscription services.
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@sugardaddyplatform.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.sugardaddyplatform.com/v1
    description: Production server
  - url: https://staging-api.sugardaddyplatform.com/v1
    description: Staging server
  - url: http://localhost:3001/v1
    description: Development server
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
        email:
          type: string
          format: email
          description: User email address
        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: '^[a-zA-Z0-9_]+$'
          description: Unique username
        firstName:
          type: string
          minLength: 1
          maxLength: 50
        lastName:
          type: string
          minLength: 1
          maxLength: 50
        role:
          type: string
          enum: [user, premium, admin]
          default: user
        isVerified:
          type: boolean
          default: false
        isActive:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        bio:
          type: string
          maxLength: 500
        location:
          type: string
          maxLength: 100
        age:
          type: integer
          minimum: 18
          maximum: 100
        gender:
          type: string
          enum: [male, female, non-binary, other]
        interests:
          type: array
          items:
            type: string
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        verificationStatus:
          type: string
          enum: [pending, verified, rejected]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserPreferences:
      type: object
      properties:
        ageRange:
          type: object
          properties:
            min:
              type: integer
              minimum: 18
              maximum: 65
            max:
              type: integer
              minimum: 18
              maximum: 100
        locationRadius:
          type: integer
          minimum: 1
          maximum: 1000
          description: Search radius in miles
        relationshipType:
          type: array
          items:
            type: string
            enum: [casual, serious, friendship, networking]
        lifestylePreferences:
          type: object
          properties:
            smoking:
              type: boolean
            drinking:
              type: boolean
            pets:
              type: boolean
            children:
              type: boolean
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        rememberMe:
          type: boolean
          default: false
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - username
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]{6,}$'
        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: '^[a-zA-Z0-9_]+$'
        firstName:
          type: string
          minLength: 1
          maxLength: 50
        lastName:
          type: string
          minLength: 1
          maxLength: 50
    Session:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        token:
          type: string
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          default: false
        message:
          type: string
        code:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account with email verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and create session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  message:
                    type: string
                  session:
                    $ref: '#/components/schemas/Session'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Account locked or suspended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Invalidate user session
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  message:
                    type: string
        '401':
          description: No active session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Retrieve current user profile and session info
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  user:
                    $ref: '#/components/schemas/User'
                  profile:
                    $ref: '#/components/schemas/UserProfile'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh session
      description: Refresh authentication token
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Session refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  session:
                    $ref: '#/components/schemas/Session'
        '401':
          description: Session expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...